% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{federal_holidays}
\alias{federal_holidays}
\title{Federal Holidays Dataset}
\format{
A data frame with columns:
\describe{
\item{date}{The observed date for the holiday.}
\item{date_definition}{Description of how the date is determined.}
\item{official_name}{The formal name of the holiday.}
\item{year_established}{The year the holiday was legally recognized.}
\item{details}{Additional information about the holiday's history.}
}
}
\source{
TidyTuesday Project
}
\usage{
federal_holidays
}
\description{
A dataset containing information about U.S. federal holidays.
}
\keyword{datasets}
