% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proposed_federal_holidays}
\alias{proposed_federal_holidays}
\title{Proposed Federal Holidays Dataset}
\format{
A data frame with columns:
\describe{
\item{date}{The proposed date for the holiday.}
\item{date_definition}{Description of how the date is determined.}
\item{official_name}{The proposed name of the holiday.}
\item{details}{Additional information about the holiday proposal.}
}
}
\source{
TidyTuesday Project
}
\usage{
proposed_federal_holidays
}
\description{
A dataset containing information about proposed U.S. federal holidays.
}
\keyword{datasets}
